+++ medusa/show_updater.py
@@ -250,7 +250,7 @@ class UpdateCache(db.DBConnection):
             'WHERE provider = ?',
             [indexer]
         )
-        return last_update_indexer[0][b'time'] if last_update_indexer else None
+        return last_update_indexer[0]['time'] if last_update_indexer else None
 
     def set_last_indexer_update(self, indexer):
         """Set the last update timestamp from the lastUpdate table.