+++ medusa/tv/episode.py
@@ -594,15 +594,15 @@ class Episode(TV):
             return True
         main_db_con = db.DBConnection()
         sql_results = main_db_con.select(
-            b'SELECT '
-            b'  * '
-            b'FROM '
-            b'  tv_episodes '
-            b'WHERE '
-            b'  indexer = ? '
-            b'  AND showid = ? '
-            b'  AND season = ? '
-            b'  AND episode = ?', [self.series.indexer, self.series.series_id, season, episode])
+            'SELECT '
+            '  * '
+            'FROM '
+            '  tv_episodes '
+            'WHERE '
+            '  indexer = ? '
+            '  AND showid = ? '
+            '  AND season = ? '
+            '  AND episode = ?', [self.series.indexer, self.series.series_id, season, episode])
 
         if len(sql_results) > 1:
             raise MultipleEpisodesInDatabaseException('Your DB has two records for the same series somehow.')
@@ -616,40 +616,40 @@ class Episode(TV):
             )
             return False
         else:
-            if sql_results[0][b'name']:
-                self.name = sql_results[0][b'name']
+            if sql_results[0]['name']:
+                self.name = sql_results[0]['name']
 
             self.season = season
             self.episode = episode
-            self.absolute_number = sql_results[0][b'absolute_number']
-            self.description = sql_results[0][b'description']
+            self.absolute_number = sql_results[0]['absolute_number']
+            self.description = sql_results[0]['description']
             if not self.description:
                 self.description = ''
-            if sql_results[0][b'subtitles'] and sql_results[0][b'subtitles']:
-                self.subtitles = sql_results[0][b'subtitles'].split(',')
-            self.subtitles_searchcount = sql_results[0][b'subtitles_searchcount']
-            self.subtitles_lastsearch = sql_results[0][b'subtitles_lastsearch']
-            self.airdate = date.fromordinal(int(sql_results[0][b'airdate']))
-            self.status = int(sql_results[0][b'status'] or UNSET)
-            self.quality = int(sql_results[0][b'quality'] or Quality.NA)
-            self.watched = int(sql_results[0][b'watched'])
+            if sql_results[0]['subtitles'] and sql_results[0]['subtitles']:
+                self.subtitles = sql_results[0]['subtitles'].split(',')
+            self.subtitles_searchcount = sql_results[0]['subtitles_searchcount']
+            self.subtitles_lastsearch = sql_results[0]['subtitles_lastsearch']
+            self.airdate = date.fromordinal(int(sql_results[0]['airdate']))
+            self.status = int(sql_results[0]['status'] or UNSET)
+            self.quality = int(sql_results[0]['quality'] or Quality.NA)
+            self.watched = int(sql_results[0]['watched'])
 
             # don't overwrite my location
-            if sql_results[0][b'location']:
-                self.location = os.path.normpath(sql_results[0][b'location'])
-            if sql_results[0][b'file_size']:
-                self.file_size = int(sql_results[0][b'file_size'])
+            if sql_results[0]['location']:
+                self.location = os.path.normpath(sql_results[0]['location'])
+            if sql_results[0]['file_size']:
+                self.file_size = int(sql_results[0]['file_size'])
             else:
                 self.file_size = 0
 
-            self.indexerid = int(sql_results[0][b'indexerid'])
-            self.indexer = int(sql_results[0][b'indexer'])
+            self.indexerid = int(sql_results[0]['indexerid'])
+            self.indexer = int(sql_results[0]['indexer'])
 
             xem_refresh(self.series)
 
-            self.scene_season = try_int(sql_results[0][b'scene_season'], 0)
-            self.scene_episode = try_int(sql_results[0][b'scene_episode'], 0)
-            self.scene_absolute_number = try_int(sql_results[0][b'scene_absolute_number'], 0)
+            self.scene_season = try_int(sql_results[0]['scene_season'], 0)
+            self.scene_episode = try_int(sql_results[0]['scene_episode'], 0)
+            self.scene_absolute_number = try_int(sql_results[0]['scene_absolute_number'], 0)
 
             if self.scene_absolute_number == 0:
                 self.scene_absolute_number = get_scene_absolute_numbering(
@@ -663,17 +663,17 @@ class Episode(TV):
                     self.season, self.episode
                 )
 
-            if sql_results[0][b'release_name'] is not None:
-                self.release_name = sql_results[0][b'release_name']
+            if sql_results[0]['release_name'] is not None:
+                self.release_name = sql_results[0]['release_name']
 
-            if sql_results[0][b'is_proper']:
-                self.is_proper = int(sql_results[0][b'is_proper'])
+            if sql_results[0]['is_proper']:
+                self.is_proper = int(sql_results[0]['is_proper'])
 
-            if sql_results[0][b'version']:
-                self.version = int(sql_results[0][b'version'])
+            if sql_results[0]['version']:
+                self.version = int(sql_results[0]['version'])
 
-            if sql_results[0][b'release_group'] is not None:
-                self.release_group = sql_results[0][b'release_group']
+            if sql_results[0]['release_group'] is not None:
+                self.release_group = sql_results[0]['release_group']
 
             self.loaded = True
             self.reset_dirty()
@@ -1155,10 +1155,10 @@ class Episode(TV):
                   {'id': self.series.series_id})
         main_db_con = db.DBConnection()
         main_db_con.action(
-            b'DELETE FROM tv_episodes '
-            b'WHERE showid = ?'
-            b' AND season = ?'
-            b' AND episode = ?',
+            'DELETE FROM tv_episodes '
+            'WHERE showid = ?'
+            ' AND season = ?'
+            ' AND episode = ?',
             [self.series.series_id, self.season, self.episode]
         )
         raise EpisodeDeletedException()
@@ -1173,56 +1173,56 @@ class Episode(TV):
 
             main_db_con = db.DBConnection()
             rows = main_db_con.select(
-                b'SELECT '
-                b'  episode_id, '
-                b'  subtitles '
-                b'FROM '
-                b'  tv_episodes '
-                b'WHERE '
-                b'  indexer = ?'
-                b'  AND showid = ? '
-                b'  AND season = ? '
-                b'  AND episode = ?',
+                'SELECT '
+                '  episode_id, '
+                '  subtitles '
+                'FROM '
+                '  tv_episodes '
+                'WHERE '
+                '  indexer = ?'
+                '  AND showid = ? '
+                '  AND season = ? '
+                '  AND episode = ?',
                 [self.series.indexer, self.series.series_id, self.season, self.episode])
 
             ep_id = None
             if rows:
-                ep_id = int(rows[0][b'episode_id'])
+                ep_id = int(rows[0]['episode_id'])
 
             if ep_id:
                 # use a custom update method to get the data into the DB for existing records.
                 # Multi or added subtitle or removed subtitles
-                if app.SUBTITLES_MULTI or not rows[0][b'subtitles'] or not self.subtitles:
+                if app.SUBTITLES_MULTI or not rows[0]['subtitles'] or not self.subtitles:
                     return [
-                        b'UPDATE '
-                        b'  tv_episodes '
-                        b'SET '
-                        b'  indexerid = ?, '
-                        b'  indexer = ?, '
-                        b'  name = ?, '
-                        b'  description = ?, '
-                        b'  subtitles = ?, '
-                        b'  subtitles_searchcount = ?, '
-                        b'  subtitles_lastsearch = ?, '
-                        b'  airdate = ?, '
-                        b'  hasnfo = ?, '
-                        b'  hastbn = ?, '
-                        b'  status = ?, '
-                        b'  quality = ?, '
-                        b'  location = ?, '
-                        b'  file_size = ?, '
-                        b'  release_name = ?, '
-                        b'  is_proper = ?, '
-                        b'  showid = ?, '
-                        b'  season = ?, '
-                        b'  episode = ?, '
-                        b'  absolute_number = ?, '
-                        b'  version = ?, '
-                        b'  release_group = ?, '
-                        b'  manually_searched = ?, '
-                        b'  watched = ? '
-                        b'WHERE '
-                        b'  episode_id = ?',
+                        'UPDATE '
+                        '  tv_episodes '
+                        'SET '
+                        '  indexerid = ?, '
+                        '  indexer = ?, '
+                        '  name = ?, '
+                        '  description = ?, '
+                        '  subtitles = ?, '
+                        '  subtitles_searchcount = ?, '
+                        '  subtitles_lastsearch = ?, '
+                        '  airdate = ?, '
+                        '  hasnfo = ?, '
+                        '  hastbn = ?, '
+                        '  status = ?, '
+                        '  quality = ?, '
+                        '  location = ?, '
+                        '  file_size = ?, '
+                        '  release_name = ?, '
+                        '  is_proper = ?, '
+                        '  showid = ?, '
+                        '  season = ?, '
+                        '  episode = ?, '
+                        '  absolute_number = ?, '
+                        '  version = ?, '
+                        '  release_group = ?, '
+                        '  manually_searched = ?, '
+                        '  watched = ? '
+                        'WHERE '
+                        '  episode_id = ?',
                         [self.indexerid, self.indexer, self.name, self.description, ','.join(self.subtitles),
                          self.subtitles_searchcount, self.subtitles_lastsearch, self.airdate.toordinal(), self.hasnfo,
                          self.hastbn, self.status, self.quality, self.location, self.file_size, self.release_name,
@@ -1232,34 +1232,34 @@ class Episode(TV):
                     # Don't update the subtitle language when the srt file doesn't contain the
                     # alpha2 code, keep value from subliminal
                     return [
-                        b'UPDATE '
-                        b'  tv_episodes '
-                        b'SET '
-                        b'  indexerid = ?, '
-                        b'  indexer = ?, '
-                        b'  name = ?, '
-                        b'  description = ?, '
-                        b'  subtitles_searchcount = ?, '
-                        b'  subtitles_lastsearch = ?, '
-                        b'  airdate = ?, '
-                        b'  hasnfo = ?, '
-                        b'  hastbn = ?, '
-                        b'  status = ?, '
-                        b'  quality = ?, '
-                        b'  location = ?, '
-                        b'  file_size = ?, '
-                        b'  release_name = ?, '
-                        b'  is_proper = ?, '
-                        b'  showid = ?, '
-                        b'  season = ?, '
-                        b'  episode = ?, '
-                        b'  absolute_number = ?, '
-                        b'  version = ?, '
-                        b'  release_group = ?, '
-                        b'  manually_searched = ?, '
-                        b'  watched = ? '
-                        b'WHERE '
-                        b'  episode_id = ?',
+                        'UPDATE '
+                        '  tv_episodes '
+                        'SET '
+                        '  indexerid = ?, '
+                        '  indexer = ?, '
+                        '  name = ?, '
+                        '  description = ?, '
+                        '  subtitles_searchcount = ?, '
+                        '  subtitles_lastsearch = ?, '
+                        '  airdate = ?, '
+                        '  hasnfo = ?, '
+                        '  hastbn = ?, '
+                        '  status = ?, '
+                        '  quality = ?, '
+                        '  location = ?, '
+                        '  file_size = ?, '
+                        '  release_name = ?, '
+                        '  is_proper = ?, '
+                        '  showid = ?, '
+                        '  season = ?, '
+                        '  episode = ?, '
+                        '  absolute_number = ?, '
+                        '  version = ?, '
+                        '  release_group = ?, '
+                        '  manually_searched = ?, '
+                        '  watched = ? '
+                        'WHERE '
+                        '  episode_id = ?',
                         [self.indexerid, self.indexer, self.name, self.description,
                          self.subtitles_searchcount, self.subtitles_lastsearch, self.airdate.toordinal(), self.hasnfo,
                          self.hastbn, self.status, self.quality, self.location, self.file_size, self.release_name,
@@ -1268,36 +1268,36 @@ class Episode(TV):
             else:
                 # use a custom insert method to get the data into the DB.
                 return [
-                    b'INSERT OR IGNORE INTO '
-                    b'  tv_episodes '
-                    b'  (episode_id, '
-                    b'  indexerid, '
-                    b'  indexer, '
-                    b'  name, '
-                    b'  description, '
-                    b'  subtitles, '
-                    b'  subtitles_searchcount, '
-                    b'  subtitles_lastsearch, '
-                    b'  airdate, '
-                    b'  hasnfo, '
-                    b'  hastbn, '
-                    b'  status, '
-                    b'  quality, '
-                    b'  location, '
-                    b'  file_size, '
-                    b'  release_name, '
-                    b'  is_proper, '
-                    b'  showid, '
-                    b'  season, '
-                    b'  episode, '
-                    b'  absolute_number, '
-                    b'  version, '
-                    b'  release_group, '
-                    b'  manually_searched, '
-                    b'  watched) '
-                    b'VALUES '
-                    b'  ((SELECT episode_id FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?), '
-                    b'  ?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);',
+                    'INSERT OR IGNORE INTO '
+                    '  tv_episodes '
+                    '  (episode_id, '
+                    '  indexerid, '
+                    '  indexer, '
+                    '  name, '
+                    '  description, '
+                    '  subtitles, '
+                    '  subtitles_searchcount, '
+                    '  subtitles_lastsearch, '
+                    '  airdate, '
+                    '  hasnfo, '
+                    '  hastbn, '
+                    '  status, '
+                    '  quality, '
+                    '  location, '
+                    '  file_size, '
+                    '  release_name, '
+                    '  is_proper, '
+                    '  showid, '
+                    '  season, '
+                    '  episode, '
+                    '  absolute_number, '
+                    '  version, '
+                    '  release_group, '
+                    '  manually_searched, '
+                    '  watched) '
+                    'VALUES '
+                    '  ((SELECT episode_id FROM tv_episodes WHERE indexer = ? AND showid = ? AND season = ? AND episode = ?), '
+                    '  ?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);',
                     [self.series.indexer, self.series.series_id, self.season, self.episode, self.indexerid, self.series.indexer, self.name,
                      self.description, ','.join(self.subtitles), self.subtitles_searchcount, self.subtitles_lastsearch,
                      self.airdate.toordinal(), self.hasnfo, self.hastbn, self.status, self.quality, self.location,
@@ -1313,38 +1313,38 @@ class Episode(TV):
             return
 
         new_value_dict = {
-            b'indexerid': self.indexerid,
-            b'name': self.name,
-            b'description': self.description,
-            b'subtitles': ','.join(self.subtitles),
-            b'subtitles_searchcount': self.subtitles_searchcount,
-            b'subtitles_lastsearch': self.subtitles_lastsearch,
-            b'airdate': self.airdate.toordinal(),
-            b'hasnfo': self.hasnfo,
-            b'hastbn': self.hastbn,
-            b'status': self.status,
-            b'quality': self.quality,
-            b'location': self.location,
-            b'file_size': self.file_size,
-            b'release_name': self.release_name,
-            b'is_proper': self.is_proper,
-            b'absolute_number': self.absolute_number,
-            b'version': self.version,
-            b'release_group': self.release_group,
-            b'manually_searched': self.manually_searched,
-            b'watched': self.watched,
+            'indexerid': self.indexerid,
+            'name': self.name,
+            'description': self.description,
+            'subtitles': ','.join(self.subtitles),
+            'subtitles_searchcount': self.subtitles_searchcount,
+            'subtitles_lastsearch': self.subtitles_lastsearch,
+            'airdate': self.airdate.toordinal(),
+            'hasnfo': self.hasnfo,
+            'hastbn': self.hastbn,
+            'status': self.status,
+            'quality': self.quality,
+            'location': self.location,
+            'file_size': self.file_size,
+            'release_name': self.release_name,
+            'is_proper': self.is_proper,
+            'absolute_number': self.absolute_number,
+            'version': self.version,
+            'release_group': self.release_group,
+            'manually_searched': self.manually_searched,
+            'watched': self.watched,
         }
 
         control_value_dict = {
-            b'indexer': self.series.indexer,
-            b'showid': self.series.series_id,
-            b'season': self.season,
-            b'episode': self.episode,
+            'indexer': self.series.indexer,
+            'showid': self.series.series_id,
+            'season': self.season,
+            'episode': self.episode,
         }
 
         # use a custom update/insert method to get the data into the DB
         main_db_con = db.DBConnection()
-        main_db_con.upsert(b'tv_episodes', new_value_dict, control_value_dict)
+        main_db_con.upsert('tv_episodes', new_value_dict, control_value_dict)
         self.loaded = False
         self.reset_dirty()